Add config file text input box
This patch adds the ses config yml file input box to the
data_bag and the UI page.  This still has no backend support
to do anything, but it shows how to correctly add a UI element
for the SES integration.

TODO: accept the input from the yaml file, process it into the
individual components and populate the other fields in the data_bag.
Then have the proposal populate all of the other fields for saving.