Fixed the ses_service object
This patch fixes a major issue with the definition of the
ses_service object.  The class was ended early and the rest
of the code in the file was ignored.  This enables the
ses_service object to work properly, which is needed to
set the new proposal defaults.

Also fixed the copyright