SES proposals can now apply
This patch fixes the ability to hit the apply button
from the UI.  It creates a dummy ses-controller chef role
that does nothing.  This also removes the dependencies on the other
barclamps.

This patch also adds a call from the cinder ceph recipe to the
ses recipe to create a ceph.conf.  Unfortunately, it still doesn't
work.