Update Nova to work with SES.
This patch changes how the nova barclamp/chef code works to support
the SES barlcamp.  First we broke up the existing ceph.rb recipe, so
that we always ensure virsh has a chance to get installed.  This is required
so that the virsh uuid secret that's shared with cinder can be created.
Also, this patch fixes an existing issue with ceph support with nova.
The existing nova.conf [libvirt] section never had support for setting
the rbd_user and rbd_secret_uuid.  I'm not sure how nova ever worked
with ceph cinder volumes without those 2 settings.

The patch calls the SES barclamp to create the required files:
/etc/ceph/ceph.conf and keyring files.